0.71 devel-12

+++ dvdisaster/debug.c
@@ -860,7 +860,7 @@ void SendCDB(char *cdb_raw)
 
 Bitmap* SimulateDefects(gint64 size)
 {  Bitmap *bm = CreateBitmap0(size);
-   int defects = (size*Closure->simulateDefects)/100;
+   gint64 defects = (size*(gint64)Closure->simulateDefects)/(gint64)100;
    
    SRandom(Closure->randomSeed);
